[chore] move folders and files.
