+++ api-payable/src/integrations/payables/payable.module.ts
@@ -3,7 +3,7 @@ import { PayableController } from './payable.controller';
 import { PayableService } from './payable.service';
 import { InfrastructureModule } from 'src/infrastructure/infrastructure.module';
 import { RabbitMQModule } from 'src/microservices/rmq/rabbitmq.module';
-import { EmailModule } from 'src/microservices/email/email.module';
+import { EmailModule } from 'src/services/email/email.module';
 
 @Module({
   imports: [
