+++ infrastructure/dev/docker/.aws-local.env
@@ -1,0 +1,10 @@
+DEBUG=${DEBUG-1}
+DOCKER_HOST=unix:///var/run/docker.sock
+AWS_DEFAULT_REGION=us-east-1
+EDGE_PORT=4566
+
+SERVICES=ecs,ses
+
+AWS_ACCESS_KEY_ID=aws_payable_key_id
+AWS_SECRET_ACCESS_KEY=aws_payable_secret_key
+AWS_S3_ENDPOINT=localhost:4572
