+++ infrastructure/dev/docker/README.md
@@ -1,0 +1,11 @@
+# Development docker
+Docker compose to create help with the development environment.
+
+## RabbitMQ
+- Management UI: `localhost:15672`.
+- Dev user login:
+  - User:     payable_user
+  - Password: payable_password
+- Guest login:
+  - User:     guest
+  - Password: guest
