+++ infrastructure/staging/docker/.api.env
@@ -3,3 +3,10 @@ NODE_ENV=production
 
 # DOCKER_SERVICE_NAME_PG=db-payable-service
 DATABASE_URL="postgres://payable_user:payable_password@db-payable-service:5432/payable_db?schema=public"
+
+RABBITMQ_HOST=localhost
+RABBITMQ_PORT=5672
+RABBITMQ_USER=payable_user
+RABBITMQ_PASSWORD=payable_password
+
+ADMIN_TEAM_EMAIL=test@email.com
