+++ infrastructure/staging/docker/.dockerignore
